+++ packages/hydrogen/src/foundation/constants.ts
@@ -1,4 +1,3 @@
 // Note: do not mix this export with other app-only logic
 // to avoid importing unnecessary code in the plugins.
-export const DEFAULT_API_VERSION = 'unstable';
 export const DEFAULT_LOCALE = 'en-us';